Bypasses cloudflares bot detection
I changed the bot to use the webdriver from the 'undetected_chromedriver' pip package which bypasses cloudflare's bot detection. When ran in headless mode, cloudflare would be able to detect it so instead of running in headless mode, it uses the 'pyvirtualdisplay' pip package to make a invisible virtual display using xvfb and then the browser runs in that.